Adds the random information content
This is the second main panel (aside from header and footer) for the
webpage. Items are inside flexbox containers for adjustability. Game
plan for this was to create an upper and lower panel; with the lower
panel containing another flexbox container that once again contains
different containers for the image and content.